old finished(?) unity build
my files for this game are all out of sorts, and the code was written while I was pretty fresh into c#, and wasn't taught things like OOP concepts yet, so i deleted all of it and restarted.
possibly the last iteration of this script before the total rewrite. there were other scripts but they were even less impressive.